Fix several issues with symbol table in Python frontend
This commit introduces custom implementation of symbol table instead of
the built-in one. It is rather primitive as it only implements API
needed by the frontend. The implementation is also very naive (may be
optimized in the future). It certainly lacks support for some Python
syntax, but such cases are presumably not supported by the frontend
analysis anyway.

The main issues with built-in symbol table:

* It cannot be generated from AST, the textual source code must be used.
* It does not differentiate between lambdas in the scope. Our
  implementation uses location information to identify a particular
  lambda.
* It does not create unique instances of Symbol class if referred to the
  same variable in all cases. This is the whole point why we use a
  symbol table instead of raw strings.

The last issue was the ultimate blocker, and so the custom
implementation must have been made.