Use single 8-byte identifier for statement id in core
Usage of file ids along with statement ids is for allowing frontends to
process multiple files independently. However, in core we collect all
the data so we are able to create unique numeric identifiers for all
statements in the program. Actual value of the identifier does not
really matter, only what matters is its uniqueness.