Use smaller integer types for pointers
Use u32 as the pointer size for arena::P, arena::S, ValueRef and StmtId.
Checks were added so if the representation is not sufficient in the
production, it will crash and we can get notified. This reduced the size
of TraceItem from 24 bytes to 8 bytes, which should help to fit it into
memory.